+++ NEWS.md
@@ -50,6 +50,7 @@
 - `drake_envir()` now throws an error, not a warning, if called in the incorrect context. Should be called only inside commands in the user's `drake` plan.
 - Replace `*expr*()` `rlang` functions with their `*quo*()` counterparts. We still keep `rlang::expr()` in the few places where we know the expressions need to be evaluated in `config$eval`.
 - The `prework` argument to `make()` and `drake_config()` can now be an expression (language object) or list of expressions. Character vectors are still acceptable.
+- At the end of `make()`, print messages about triggers etc. only if `verbose >= 2L`.
 
 
 # Version 6.2.1